Create infrastructure to update reward redemtpion status
This change provides everything that's needed to automatically update the status
of a reward redemption and triggers it based on whether a color reward was
redeemed with the correct input (i.e. a parseable color) or not. Unfortunately,
it can't be activated because the Twitch API only allows updating the status of
rewards that were created by the same client.

Thanks Pasqual/derklaro/einklaro for the PR! #4 😃